Initial register map for WM8960
Because WM8960 does not provide a read interface the register map is
typically kept in memory by a driver and writes sent over I2C to the
codec as needed. This allows for partial register updates as the state
of the device is always kept in memory.

Re-use of chiptools yaml descriptor format is helpful here for
describing registers and fields though in the end a custom chiptool IR
to Rust generator will likely be needed.

The reset value of all registers is used as the initial value of all
registers. Otherwise a map is used along with i2c for writes to
manipulate registers.

The generated files are cached but can be rebuilt by enableing the
codegen feature. This is because the chiptool version required is
an unreleased fork.